tools/example-mypy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "example-mypy"
dynamic = ["version"]
license = { text = "MIT" }
authors = [
    { name = "Stephen Moore", email = "stephen@delfick.com" },
]
dependencies = [
    "extended-mypy-django-plugin==0.7.2",
    "mypy==1.11.2",
    "django-stubs==5.1.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "example_mypy/version.py"

[tool.hatch.build.targets.wheel]
include = [
    "/example_mypy",
]
